* dev: (107 commits)
  fix display of activity alerts when org info is missing
  updated AlertLoader
  fix issue with processing org activity alerts
  show validation error when copying linked workflows across Org
  support workflow activity alerts
  updated dependency of useEffect to avoid double call
  automation task in workflow builder is only visible to dock pro users
  show patient labels in filter options
  Fix: Prevent duplicate profile creation during rename
  Use dateTimeIntent for Workflow
  added PATIENT_LABELS in filter Option
  Used MarkdownRenderer Component in AI summary Modal
  Created MarkdownRenderer Component
  Refactor: Replace string literals with DueDateIntent constants and clean up unused code
  Use dateTimeIntent for Task
  added workflows in filter options Category
  Fix: due date display issue in Task List
  updated storage keys
  updated the condition for edit mode
  corrected typo
  ...

+++ src/app/views/patient-details/PatientAttachments/hooks.js
@@ -141,15 +141,39 @@ const useInitializeAttachmentsSectionHooks = () => {
 
   const deleteAttachment = useCallback(
     (identifier) => {
-      dispatch(
-        PatientDetailsActions.deletePatientAttachment(
-          patientIdentifier,
-          identifier,
-        ),
-      );
+      dispatch(openModal('DeleteConfirmation',{
+        title: 'Delete Attachment',
+        description: 'Are you sure you want to delete this attachment? This action cannot be undone.',
+        confirm: () => {
+          dispatch(
+            PatientDetailsActions.deletePatientAttachment(
+              patientIdentifier,
+              identifier,
+            ),
+          );
+        }
+      }))
     },
     [dispatch, patientIdentifier],
   );
+
+  const downloadAttachment = async (attachment) => {
+    const { attachmentIdentifier, fileName, contentType } = attachment;
+    try {
+        const {data} = await getMemoPatientAttachment(attachmentIdentifier);
+        const blob = new Blob([data], { type: contentType });
+        const url = window.URL.createObjectURL(blob);
+        const anchor = document.createElement('a');
+        anchor.href = url;
+        anchor.download = fileName;
+        document.body.appendChild(anchor);
+        anchor.click();
+        document.body.removeChild(anchor);
+        window.URL.revokeObjectURL(url);
+    } catch (error) {
+        console.error('Error downloading attachment:', error);
+    } 
+};
 
   const openAttachmentPreview = useCallback(
     (attachment) => {
@@ -313,6 +337,7 @@ const useInitializeAttachmentsSectionHooks = () => {
       isDragActive,
     },
     downloadAllFiles,
+    downloadAttachment,
     getMemoPatientAttachment,
   };
 };
